Arian, would be great if you could look closer at this new mixin to see if it conflicts with the remotebrowser's download functionality.
Also, I didn't manage to get this rather simple example to work, it times out. Any ideas why:

client = BrowserbaseBrowser(
    dendrite_api_key=os.getenv("DENDRITE_API_KEY"),
    openai_api_key=os.getenv("OPENAI_API_KEY"),
    anthropic_api_key=os.getenv("ANTHROPIC_API_KEY"),
    enable_downloads=True,
)
client.goto(
    "https://github.com/dendrite-systems/dendrite-python-sdk/blob/main/scripts/generate_sync.py"
)
client.click("download raw button")
download = client.get_download()
print(download.save_as("generate_sync.py"))
client.close()
