+++ dendrite_sdk/sync_api/_core/_base_browser.py
@@ -24,6 +24,7 @@ from dendrite_sdk.sync_api._core.mixin.extract import ExtractionMixin
 from dendrite_sdk.sync_api._core.mixin.fill_fields import FillFieldsMixin
 from dendrite_sdk.sync_api._core.mixin.get_element import GetElementMixin
 from dendrite_sdk.sync_api._core.mixin.keyboard import KeyboardMixin
+from dendrite_sdk.sync_api._core.mixin.upload_download_mixin import DownloadUploadMixin
 from dendrite_sdk.sync_api._core.mixin.wait_for import WaitForMixin
 from dendrite_sdk.sync_api._core.models.authentication import AuthSession
 from dendrite_sdk.sync_api._core.models.api_config import APIConfig
@@ -36,6 +37,7 @@ from dendrite_sdk._common._exceptions.dendrite_exception import (
 
 
 class BaseDendrite(
+    DownloadUploadMixin,
     ExtractionMixin,
     WaitForMixin,
     AskMixin,
